requirements.txt
transformers
torch
Pillow
datasets
opencv-python
gradio
numpy
